requirements.txt

aiogram==3.4.1
sqlalchemy==2.0.23
psycopg[binary]==3.1.12
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
beautifulsoup4==4.12.2
